Move ImageActivity to common package
